refactor: inline singleton hooks
